Refactor: memberenum -> type 으로 패키지명 변경 (#21)
