+++ src/main/java/org/sesac/slopedbe/member/model/type/MemberRole.java
@@ -1,13 +1,14 @@
-package org.sesac.slopedbe.member.model.memberenum;
+package org.sesac.slopedbe.member.model.type;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum MemberOauthType {
-	NAVER("TYPE_NAVER"),
-	GOOGLE("TYPE_GOOGLE");
+public enum MemberRole {
+	ADMIN("ROLE_ADMIN"),
+	USER("ROLE_USER");
+
 
 	private final String value;
 }
